refactor: updated root cmd to display version with -v or --version flag. template updated.
