--- .version.yaml
@@ -1,6 +1,0 @@
-alpha: 0
-beta: 0
-major: 1
-minor: 0
-patch: 1
-rc: 0
